move code to keep screen on from homeactivity to base fragment rename 'Mix' to 'Mixed' in transactions details dialog add nav bar to license page
