+++ app/src/main/java/com/dcrandroid/HomeActivity.kt
@@ -21,7 +21,6 @@ import android.os.Handler
 import android.util.DisplayMetrics
 import android.view.MotionEvent
 import android.view.View
-import android.view.WindowManager
 import android.view.animation.AnimationUtils
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
@@ -58,7 +57,7 @@ import kotlin.system.exitProcess
 
 const val TAG = "HomeActivity"
 
-class HomeActivity : BaseActivity(), SyncProgressListener, TxAndBlockNotificationListener, ProposalNotificationListener, AccountMixerNotificationListener {
+class HomeActivity : BaseActivity(), SyncProgressListener, TxAndBlockNotificationListener, ProposalNotificationListener {
 
     private var deviceWidth: Int = 0
     private var blockNotificationSound: Int = 0
@@ -154,8 +153,6 @@ class HomeActivity : BaseActivity(), SyncProgressListener, TxAndBlockNotificatio
                 e.printStackTrace()
             }
         }
-
-        checkMixerStatus()
     }
 
     private val bottomSheetDismissed = DialogInterface.OnDismissListener {
@@ -177,32 +174,8 @@ class HomeActivity : BaseActivity(), SyncProgressListener, TxAndBlockNotificatio
         }
     }
 
-    private fun checkMixerStatus() = GlobalScope.launch(Dispatchers.Main) {
-        var activeMixers = 0
-        for (wallet in multiWallet!!.openedWalletsList()) {
-            if (wallet.isAccountMixerActive) {
-                activeMixers++
-            }
-        }
-
-        if (activeMixers > 0) {
-            window.addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
-        } else {
-            window.clearFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
-        }
-    }
-
-    override fun onAccountMixerEnded(walletID: Long) {
-        checkMixerStatus()
-    }
-
-    override fun onAccountMixerStarted(walletID: Long) {
-        checkMixerStatus()
-    }
-
     override fun onResume() {
         super.onResume()
-        checkMixerStatus()
     }
 
     override fun onDestroy() {
