+++ app/src/main/java/com/dcrandroid/fragments/BaseFragment.kt
@@ -6,15 +6,20 @@
 
 package com.dcrandroid.fragments
 
+import android.view.WindowManager
 import androidx.annotation.StringRes
 import androidx.fragment.app.Fragment
 import com.dcrandroid.HomeActivity
 import com.dcrandroid.data.Transaction
+import com.dcrandroid.extensions.openedWalletsList
 import com.dcrandroid.util.WalletData
 import com.google.gson.Gson
 import dcrlibwallet.*
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.launch
 
-open class BaseFragment : Fragment(), SyncProgressListener, TxAndBlockNotificationListener {
+open class BaseFragment : Fragment(), SyncProgressListener, TxAndBlockNotificationListener, AccountMixerNotificationListener {
 
     var TAG = this.javaClass.name
     var isForeground = false
@@ -31,6 +36,8 @@ open class BaseFragment : Fragment(), SyncProgressListener, TxAndBlockNotificati
 
         multiWallet?.addSyncProgressListener(this, TAG)
         multiWallet?.addTxAndBlockNotificationListener(this, TAG)
+
+        checkMixerStatus()
     }
 
     override fun onResume() {
@@ -40,6 +47,7 @@ open class BaseFragment : Fragment(), SyncProgressListener, TxAndBlockNotificati
             requiresDataUpdate = false
             onTxOrBalanceUpdateRequired(null)
         }
+        checkMixerStatus()
     }
 
     override fun onPause() {
@@ -51,6 +59,29 @@ open class BaseFragment : Fragment(), SyncProgressListener, TxAndBlockNotificati
         super.onDestroy()
         multiWallet?.removeSyncProgressListener(TAG)
         multiWallet?.removeTxAndBlockNotificationListener(TAG)
+    }
+
+    private fun checkMixerStatus() = GlobalScope.launch(Dispatchers.Main) {
+        var activeMixers = 0
+        for (wallet in multiWallet!!.openedWalletsList()) {
+            if (wallet.isAccountMixerActive) {
+                activeMixers++
+            }
+        }
+
+        if (activeMixers > 0) {
+            requireActivity().window.addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
+        } else {
+            requireActivity().window.clearFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
+        }
+    }
+
+    override fun onAccountMixerEnded(walletID: Long) {
+        checkMixerStatus()
+    }
+
+    override fun onAccountMixerStarted(walletID: Long) {
+        checkMixerStatus()
     }
 
     fun setToolbarTitle(title: CharSequence, showShadow: Boolean) {
